+++ src/main/java/com/cpa/yusin/quiz/bookmark/controller/port/GetBookmarkStatusService.java
@@ -1,0 +1,9 @@
+package com.cpa.yusin.quiz.bookmark.controller.port;
+
+import com.cpa.yusin.quiz.bookmark.controller.dto.request.BookmarkStatusRequest;
+import com.cpa.yusin.quiz.bookmark.controller.dto.response.BookmarkStatusResponse;
+
+public interface GetBookmarkStatusService {
+
+    BookmarkStatusResponse getBookmarkStatus(Long memberId, BookmarkStatusRequest request);
+}